core/tostr: Moved ofi_strncatf() and ofi_nic_tostr() to ofi_str.c.
Moved functions ofi_strncatf(), ofi_nic_tostr() and other static
ofi_tostr_xxx functions from common.c to ofi_str.c. The declaration of
these functions are moved from ofi.h to ofi_str.h.
